0.1.1: url-encode javascript code
This fixes bugs resulting from minified JS incidentally containing sequences of characters that happen to look like URL-encoded data, which the browser decodes into incorrect JavaScript.

As it turns out, URL-encoding is all we actually need to do. Instead of stripping out line comments and newlines and escaping double-quotes, we can just URL-encode the whole dang thing and it seems to work fine.